特效相机完成预览部分，和录制框架搭建
