+++ EffectLib/src/main/java/com/gpufast/effect/filter/OesToRgbFilter.java
@@ -3,7 +3,7 @@ package com.gpufast.effect.filter;
 import android.opengl.GLES11Ext;
 import android.opengl.GLES20;
 
-import com.gpufast.gles.GlUtil;
+import com.gpufast.gles.GLESUtil;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
@@ -100,7 +100,7 @@ public class OesToRgbFilter {
     public void init() {
 
 
-        mProgram = GlUtil.createProgram(VERTEX_SHADER,FRAGMENT_SHADER);
+        mProgram = GLESUtil.createProgram(VERTEX_SHADER,FRAGMENT_SHADER);
         GLES20.glUseProgram(mProgram);
 
         positionLoc = GLES20.glGetAttribLocation(mProgram, "vPosition");
@@ -124,14 +124,14 @@ public class OesToRgbFilter {
      * 准备离屏缓冲区
      */
     private void prepareFramebuffer(int width, int height) {
-        GlUtil.checkGlError("prepareFramebuffer start");
+        GLESUtil.checkGlError("prepareFramebuffer start");
         int[] values = new int[1];
         // 创建颜色缓冲区
         GLES20.glGenTextures(1, values, 0);
-        GlUtil.checkGlError("glGenTextures");
+        GLESUtil.checkGlError("glGenTextures");
         mFrameBufferTextureId = values[0];   // expected > 0
         GLES20.glBindTexture(GLES20.GL_TEXTURE_2D, mFrameBufferTextureId);
-        GlUtil.checkGlError("glBindTexture " + mFrameBufferTextureId);
+        GLESUtil.checkGlError("glBindTexture " + mFrameBufferTextureId);
 
         // 设置texture存储
         GLES20.glTexImage2D(GLES20.GL_TEXTURE_2D, 0, GLES20.GL_RGBA, width, height, 0,
@@ -146,20 +146,20 @@ public class OesToRgbFilter {
                 GLES20.GL_CLAMP_TO_EDGE);
         GLES20.glTexParameteri(GLES20.GL_TEXTURE_2D, GLES20.GL_TEXTURE_WRAP_T,
                 GLES20.GL_CLAMP_TO_EDGE);
-        GlUtil.checkGlError("glTexParameter");
+        GLESUtil.checkGlError("glTexParameter");
 
         // 创建帧缓冲区
         GLES20.glGenFramebuffers(1, values, 0);
-        GlUtil.checkGlError("glGenFramebuffers");
+        GLESUtil.checkGlError("glGenFramebuffers");
         mFramebuffer = values[0];
         GLES20.glBindFramebuffer(GLES20.GL_FRAMEBUFFER, mFramebuffer);
-        GlUtil.checkGlError("glBindFramebuffer " + mFramebuffer);
+        GLESUtil.checkGlError("glBindFramebuffer " + mFramebuffer);
 
 
         //将颜色缓冲区绑定到帧缓冲区中
         GLES20.glFramebufferTexture2D(GLES20.GL_FRAMEBUFFER, GLES20.GL_COLOR_ATTACHMENT0,
                 GLES20.GL_TEXTURE_2D, mFrameBufferTextureId, 0);
-        GlUtil.checkGlError("glFramebufferTexture2D");
+        GLESUtil.checkGlError("glFramebufferTexture2D");
 
         // See if GLES is happy with all this.
         int status = GLES20.glCheckFramebufferStatus(GLES20.GL_FRAMEBUFFER);
@@ -170,7 +170,7 @@ public class OesToRgbFilter {
         //切换回默认的缓冲区
         GLES20.glBindFramebuffer(GLES20.GL_FRAMEBUFFER, 0);
 
-        GlUtil.checkGlError("prepareFramebuffer done");
+        GLESUtil.checkGlError("prepareFramebuffer done");
     }
